Dev and prod operational
Both environments are now operational - the development will eventually use CORS (to enable some authentication options) and is featuring server restart. Production is serving static content that is built from create-react-app in a parallel folder. The naming of folders is hardcoded so if it's not backend and frontend, the scripts should reflect the naming changes.